% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cause_of_death}
\alias{cause_of_death}
\title{Cause of death code}
\format{
\describe{
\item{code}{Index}
\item{description_en}{Description in English}
\item{description_no}{Description in Norwegian}
\item{code_icd10}{ICD-10 code range}
}
}
\usage{
cause_of_death
}
\description{
This dataset contains the European shortlist for cause of death (English, Norwegian)
}
\keyword{datasets}
