% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findcode.R
\name{find_code}
\alias{find_code}
\title{Find description given an ICD10 code}
\usage{
find_code(code, d = medicode::icd10_en)
}
\arguments{
\item{code}{A c}

\item{d}{Dataset that we search. Default is the english table}
}
\value{
A data.table with code and description
}
\description{
Find description given an ICD10 code
}
\examples{
find_code(code = 'A001')
}
