% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{icd10_en}
\alias{icd10_en}
\title{ICD-10 code in English}
\format{
\describe{
\item{code}{ICD-10 code}
\item{description_en}{Description in English}
}
}
\usage{
icd10_en
}
\description{
This dataset contains the ICD-10 code in English.
}
\keyword{datasets}
