% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{icd10_no}
\alias{icd10_no}
\title{ICD-10 code in Norwegian}
\format{
\describe{
\item{code}{ICD-10 code}
\item{description_no}{Description in Norwegian}
}
}
\usage{
icd10_no
}
\description{
This dataset contains the ICD-10 code in Norwegian.
}
\keyword{datasets}
